requirements.txt
absl-py==1.4.0
appdirs==1.4.4
cachetools==5.3.1
certifi==2023.5.7
cffi==1.16.0
charset-normalizer==3.2.0
click==8.1.5
cloudpickle==2.2.1
cmake==3.26.4
contourpy==1.1.1
cycler==0.12.1
Cython==0.29.36
docker-pycreds==0.4.0
Farama-Notifications==0.0.4
fasteners==0.19
filelock==3.12.2
fonttools==4.43.1
fsspec==2023.9.2
gitdb==4.0.10
GitPython==3.1.32
glfw==2.6.2
google-auth==2.22.0
google-auth-oauthlib==1.0.0
grpcio==1.56.0
gym==0.21.0
gymnasium==0.29.0
idna==3.4
imageio==2.31.5
importlib-metadata==6.8.0
Jinja2==3.1.2
joblib==1.3.1
kiwisolver==1.4.5
lit==16.0.6
llvmlite==0.41.0
locket==1.0.0
Markdown==3.4.3
MarkupSafe==2.1.3
mpmath==1.3.0
msgpack==1.0.7
mujoco-py==2.1.2.14
networkx==3.1
numba==0.58.0
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
packaging==23.1
pandas==2.0.3
partd==1.4.1
pathtools==0.1.2
patsy==0.5.3
Pillow==10.0.1
protobuf==4.23.4
psutil==5.9.5
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.3.0
sentry-sdk==1.28.1
setproctitle==1.3.2
six==1.16.0
smmap==5.0.0
sortedcontainers==2.4.0
stable-baselines3==1.5.0
statsmodels==0.14.0
stumpy==1.12.0
sympy==1.12
tblib==2.0.0
tensorboard==2.13.0
tensorboard-data-server==0.7.1
threadpoolctl==3.2.0
toolz==0.12.0
torch==2.0.1
tornado==6.3.3
tqdm==4.66.1
triton==2.0.0
tsfresh==0.20.1
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.16
wandb==0.15.5
Werkzeug==2.3.6
zict==3.0.0
zipp==3.16.2
rpy2